requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
-e git+git://github.com/JohanWieslander/garmin-uploader@master#egg=garmin-uploader  # via -r requirements/base.in
amqp==5.0.1               # via kombu
appdirs==1.4.4            # via virtualenv
arrow==0.17.0             # via stravalib
attrs==19.3.0             # via codaio
backcall==0.2.0           # via ipython
bcrypt==3.2.0             # via paramiko
billiard==3.6.3.0         # via celery
celery==5.0.1             # via -r requirements/base.in, flower
certifi==2020.6.20        # via requests, sentry-sdk
cffi==1.14.3              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click-didyoumean==0.0.3   # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-repl, pip-tools
codaio==0.6.3             # via -r requirements/base.in
cryptography==3.2.1       # via paramiko, social-auth-core
decorator==4.4.2          # via codaio, ipython
defusedxml==0.6.0         # via python3-openid, social-auth-core
distlib==0.3.1            # via virtualenv
dj-database-url==0.5.0    # via -r requirements/base.in
django-appconf==1.0.4     # via -r requirements/dev.in
django-debug-toolbar==3.1.1  # via -r requirements/dev.in
django-extensions==3.0.9  # via -r requirements/dev.in
django-geojson==3.1.0     # via -r requirements/base.in
django-leaflet==0.27.1    # via -r requirements/base.in
django-widget-tweaks==1.4.8  # via -r requirements/base.in
django==2.2.17            # via -r requirements/base.in, django-appconf, django-debug-toolbar, django-geojson, django-leaflet, easy-thumbnails
easy-thumbnails==2.7      # via -r requirements/base.in
envparse==0.2.0           # via codaio
fabric3==1.14.post1       # via -r requirements/dev.in
filelock==3.0.12          # via tox, virtualenv
flower==0.9.5             # via -r requirements/base.in
future==0.18.2            # via -r requirements/base.in
gitric==0.4               # via -r requirements/dev.in
gpxpy==1.4.2              # via -r requirements/base.in
gunicorn==20.0.4          # via -r requirements/base.in
humanize==3.1.0           # via flower
idna==2.10                # via requests
importlib-metadata==2.0.0  # via kombu, pluggy, tox, virtualenv
inflection==0.3.1         # via codaio
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements/dev.in
jedi==0.17.2              # via ipython
joblib==0.17.0            # via scikit-learn
kombu==5.0.2              # via celery
lxml==4.6.1               # via -r requirements/base.in
numexpr==2.7.1            # via tables
numpy==1.19.3             # via numexpr, pandas, scikit-learn, scipy, tables
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
packaging==20.4           # via tox
pandas==1.1.4             # via -r requirements/base.in
paramiko==2.7.2           # via fabric3
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via -r requirements/base.in, easy-thumbnails
pip-tools==5.3.1          # via -r requirements/dev.in
pluggy==0.13.1            # via tox
polyline==1.4.0           # via -r requirements/base.in
prometheus-client==0.8.0  # via flower
prompt-toolkit==3.0.8     # via click-repl, ipython
psycopg2-binary==2.8.6    # via -r requirements/base.in
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via tox
pycparser==2.20           # via cffi
pygments==2.7.2           # via ipython
pyjwt==1.7.1              # via social-auth-core
pynacl==1.4.0             # via paramiko
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via arrow, codaio, pandas
python3-openid==3.2.0     # via social-auth-core
pytz==2020.4              # via celery, django, flower, pandas, stravalib
rcssmin==1.0.6            # via -r requirements/base.in
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.24.0          # via -r requirements/base.in, codaio, garmin-uploader, requests-oauthlib, social-auth-core, stravalib
scikit-learn==0.23.2      # via -r requirements/base.in
scipy==1.5.3              # via scikit-learn
sentry-sdk==0.19.1        # via -r requirements/base.in
six==1.15.0               # via bcrypt, click-repl, cryptography, fabric3, garmin-uploader, packaging, pip-tools, polyline, pynacl, python-dateutil, social-auth-app-django, social-auth-core, stravalib, tox, virtualenv
social-auth-app-django==4.0.0  # via -r requirements/base.in
social-auth-core==3.3.3   # via -r requirements/base.in, social-auth-app-django
sqlparse==0.4.1           # via django, django-debug-toolbar
stravalib==0.10.2         # via -r requirements/base.in
tables==3.6.1             # via -r requirements/base.in
threadpoolctl==2.1.0      # via scikit-learn
toml==0.10.2              # via tox
tornado==6.1              # via flower
tox==3.20.1               # via -r requirements/dev.in
tqdm==4.51.0              # via -r requirements/base.in
traitlets==5.0.5          # via ipython
units==0.7                # via stravalib
urllib3==1.25.11          # via requests, sentry-sdk
vine==5.0.0               # via amqp, celery
virtualenv==20.1.0        # via tox
wcwidth==0.2.5            # via prompt-toolkit
werkzeug==1.0.1           # via -r requirements/dev.in
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
